+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Use regex in `Parser` to translate more Doxygen commands into Javadoc tags ([pull #278](https://github.com/bytedeco/javacpp/pull/278) and [pull #281](https://github.com/bytedeco/javacpp/pull/281))
  * Do not let `Parser` map `operator=()` when prefixing container name with `const ` ([pull #280](https://github.com/bytedeco/javacpp/pull/280))
 
 ### January 11, 2019 version 1.4.4